サイトのレイアウトを変更: レトロデザインをメインに
- /oldのレトロデザインをルートページ(/)に移動
- 従来のモダンデザインを/oldに移動
- コンポーネント名とページタイトルを適切に変更
